Move pipes to pipe folder
