+++ client/src/app/pipes/index.ts
@@ -1,0 +1,2 @@
+export * from './display-seconds.pipe'
+export * from './display-track.pipe'
